AR pass 1 over #28 T1/T2a: 4 High, 7 Medium, 8 Low — the Highs fixed
The previous commit was broken and this corrects it. An adversarial review
pass proved four High findings by executing probes against the built
assemblies; three are fixed here, with the remainder tracked.

ERR-028-006 (High) — every player read as age 0 from the first daily step of
every new game. A new world starts on day 0, so the anchor
newGameDay - Age0*DAYS_PER_YEAR is negative for anyone with a non-zero
generated age; held as uint it was clamped to 0, making the derived age
worldDay/365. Measured: bootstrap ages 26,22,30,26,28,30 -> 0,0,0,0,0,0
after one day; bands over 100 players growth=100 stable=0 decline=0. So the
Decline band was unreachable, RETIREMENT_AGE could never fire, and — because
KD-4 made SquadFor the roster authority — Age = 0 was projected into
LineupSelector, SquadRating and MatchEngine.ConfigureSquads.

  BirthWorldDay is now a signed long and the codec field u32 -> i64 (free:
  the format is v1 and unreleased, so there is no migration to break). Both
  clamps deleted. Re-measured at day 0: ages unchanged through the first
  advance, bands growth=42 stable=26 decline=32.

  The suite could not have caught this. Both #28 fixtures use
  BaseDay = 100000 with a comment stating it keeps the anchor non-negative —
  written around the defect, so no test ever ran the day the product starts
  on. Recorded in #28 §5 as a fixture hazard.

  The landing's own headline lock was complicit: it asserted
  days x players == |sum of cursors|, which is satisfiable only when every
  player sits in ONE band, so it passed BECAUSE of this defect. Rewritten to
  assert each player's cursor against his bootstrap age (an independent
  source) with all-three-bands preconditions, and re-verified to kill the
  slot-1 mutant.

ERR-028-007 (High) — LastAdvancedWorldDay is the fourth persisted per-player
cursor and was checked at none of the three boundaries the #29/#41 review
loop spent passes 5, 6 and 9 establishing. A cursor 9,999 days ahead of a
clock at 0 was accepted at composition, Save and Load, and silently froze
growth. Lagging is worse here than for the siblings because AdvanceDay
replays a gap, banking N days of growth from one day's inputs. Now one
shared predicate with all three boundaries delegating.

ERR-028-008 (High) — Save's `?? ProgressionEngine.Empty` let a resume that
dropped the store write a zero-club roster over a populated file: clubs=4 ->
clubs=0, every gate green. The reviewer's proposed fix was NARROWED after
testing it: refusing career-without-progression broke four suites that build
career-wired loops over a mutable bootstrap provider, which is a legitimate
pre-#28 composition that round-trips correctly. The guard now reads the
destination and refuses only an empty block overwriting a populated one.

ERR-028-009 (Medium) — the never-advanced sentinel was a legal worldDay: it
stored (breaking idempotency) and hung the gap-replay loop forever at
uint.MaxValue. Refused up front, matching both siblings' F8 rows.

M6 — season-save-tests.asmdef never gained the PlayerProgression reference.
The dotnet gate compiles regardless because SDK ProjectReferences are
transitive and Unity asmdef references are not, so the Editor build was
broken invisibly to the only compiler this project runs. Fixed, and every
test `using` audited against its asmdef.

Suites: PlayerProgression.Tests 41 -> 70, SeasonSave.Tests 361 -> 372, all
green; solution builds 0 errors; recurring-defect-lint 0 ERROR tree-wide.
The H1 fix is mutation-verified by three independent locks.

Still open from this pass: H2 (a progression-wired SeasonLoop cannot play a
round through any public API — the constructor builds the provider and
AdvanceAndPlayNextRound demands reference-equality with an instance nothing
exposes), M2, M3, M5, M7 and the Lows. Whole-tree gate still not run.

+++ src/player-progression/PlayerLifecycle.cs
@@ -29,8 +29,17 @@ namespace TacticalDirector.PlayerProgression
         /// The authoritative age anchor (KD-A): age is DERIVED as <c>(worldDay − BirthWorldDay) /
         /// DAYS_PER_YEAR</c>, so there is no discrete year-rollover step to double-count (FR-PG-005).
         /// Pinned once at new-game from the generation-time age.
+        /// <para>
+        /// <b>Signed, and that is load-bearing (ERR-028-006).</b> §3.1.1 pins the anchor as
+        /// <c>newGameDay − Age0 · DAYS_PER_YEAR</c>, and a new world starts on day <b>0</b> — so for
+        /// every player with a non-zero generated age the anchor is NEGATIVE. Held as a <c>uint</c> it
+        /// had to be clamped to 0, which made the derived age <c>worldDay / 365</c>: the entire league
+        /// became age 0 on the first daily step, the Decline band was unreachable and
+        /// <c>RETIREMENT_AGE</c> could never fire. A player born before the epoch is an ordinary state,
+        /// not an edge case, so the anchor must be able to represent it.
+        /// </para>
         /// </summary>
-        public uint BirthWorldDay;
+        public long BirthWorldDay;
 
         /// <summary>Set on the world tick at RETIREMENT_AGE (FR-PG-013); the player stays selectable until the season boundary.</summary>
         public bool RetirementFlag;
